moved module file to proper location so cli generation works correctly
