+++ frontend/src/app/app.module.ts
@@ -5,23 +5,23 @@ import { ReactiveFormsModule, FormControl } from '@angular/forms';
 import { RouterModule, Routes } from '@angular/router';
 import { AgmCoreModule } from '@agm/core';
 
-import { AppComponent } from './app.component';
-import { AppBarComponent } from '../app-bar/app-bar.component';
-import { ChatComponent } from '../chat/chat.component';
-import { MapComponent } from '../map/map.component';
-import { LoginComponent } from '../login/login.component';
-import { RegisterComponent } from '../register/register.component';
-import { UserService } from '../../services/user-service/user.service';
-import { LocationComponent } from '../locations/location/location.component';
-import { MapService } from '../../services/map/map.service';
-import { TeamService } from '../../services/team-service/team.service';
-import { LocationDetailsComponent } from '../location-details/location-details.component';
-import { TeamsHomeComponent } from '../teams-home/teams-home.component';
-import { CreateTeamComponent } from '../teams-create/create-team.component';
-import { TeamListComponent } from '../team-list/team-list.component';
-import { UserTeamsComponent } from '../user-teams/user-teams.component';
-import { JoinTeamComponent } from '../teams-join/join-team.component';
-import { TeamMatchComponent} from "../team-match/team-match.component";
+import { AppComponent } from './components/app/app.component';
+import { AppBarComponent } from './components/app-bar/app-bar.component';
+import { ChatComponent } from './components/chat/chat.component';
+import { MapComponent } from './components/map/map.component';
+import { LoginComponent } from './components/login/login.component';
+import { RegisterComponent } from './components/register/register.component';
+import { UserService } from './services/user-service/user.service';
+import { LocationComponent } from './components/locations/location/location.component';
+import { MapService } from './services/map/map.service';
+import { TeamService } from './services/team-service/team.service';
+import { LocationDetailsComponent } from './components/location-details/location-details.component';
+import { TeamsHomeComponent } from './components/teams-home/teams-home.component';
+import { CreateTeamComponent } from './components/teams-create/create-team.component';
+import { TeamListComponent } from './components/team-list/team-list.component';
+import { UserTeamsComponent } from './components/user-teams/user-teams.component';
+import { JoinTeamComponent } from './components/teams-join/join-team.component';
+import { TeamMatchComponent} from "./components/team-match/team-match.component";
 import {FlexLayoutModule} from "@angular/flex-layout";
 import {MatButtonModule, MatFormFieldModule, MatInputModule, MatToolbarModule} from "@angular/material";
 import {BrowserAnimationsModule} from "@angular/platform-browser/animations";
