+++ frontend/src/app/sandbox/sandbox.component.html
@@ -1,0 +1,3 @@
+<p>
+  sandbox works!
+</p>
