+++ frontend/src/app/sandbox/sandbox.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-sandbox',
+  templateUrl: './sandbox.component.html',
+  styleUrls: ['./sandbox.component.css']
+})
+export class SandboxComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
